+++ docs/changes.rst
@@ -11,6 +11,7 @@ Not yet released.
 * Deprecated insecure configuration of VCS service API keys via _TOKEN/_USERNAME configuration instead of _CREDENTIALS list.
 * Fixed processing of some uploaded CSV files.
 * Improved whitespace changes handling in diff display.
+* Added automatic suggestions management link to management pages.
 
 `All changes in detail <https://github.com/WeblateOrg/weblate/milestone/87?closed=1>`__.